+++ flutter/lib/ui/history/app_bar_content.dart
